requirements.txt
streamlit
tensorflow
keras
numpy
Pillow
opencv-python
ultralytics
torch
torchvision
pandas

